Update README.md
Task 2 - Terraform and Cloud: create the infrastructure to host your container.
Using Terraform, create the following infrastructure in AWS (or equivalent):

If server-based:
A VPC with 2 public and 2 private subnets.
An ECS/EKS or equivalent cluster deployed to that VPC.
A ECS/EKS task/service resource to run your container
The tasks and/nodes must be on the private subnets only.
A load balancer deployed in the public subnets to offer the service.
If serverless:
A VPC with 2 public and 2 private subnets.
A Lambda or equivalent function running your container
Appropriate configuration to associate the function with the private subnets.
An API Gateway, CDN, or loadbalancer to trigger your function
If you prefer, you may use popular modules from the Terraform registry (for example the VPC and EKS modules).

Push your code to a public git repository
Push your code to a public git repository in the platform of your choice (e.g. GitHub, GitLab, Bitbucket, etc.). MAKE SURE YOU DON'T PUSH ANY SECRETS LIKE API KEYS TO A PUBLIC REPO!
We have a recommended repository structure here.